… .github/workflows/buildjet-release.yml
… .github/workflows/distributed-release.yml
… .goreleaser.yml
… docs/BUILD_OPTIMIZATION.md
… scripts/fix-clippy.sh
… scripts/goreleaser-pgo.sh